Create wrappers dir. Move processing to utils. Move model_zoo to models. Update README.
